+++ lib/components/Button/ButtonGroup.js
@@ -1,7 +1,7 @@
 import React, { Component, Children } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
-import DropdownButton from './../DropdownMenu';
+import DropdownButton from './../Dropdown';
 
 export default class ButtonGroup extends Component {
   constructor() {
