+++ lib/components/Lookup/Lookup.js
@@ -6,7 +6,7 @@ import Input from './../Input';
 import Icon from './../Icon';
 import Spinner from './../Spinner';
 import Pill from './../Pill';
-import { DropdownButton, DropdownMenuItem } from './../DropdownMenu';
+import { DropdownButton, DropdownMenuItem } from './../Dropdown';
 import { autoAlign, uuid, isElInChildren, registerStyle } from './../utils';
 
 /**
