+++ lib/components/Picklist/Picklist.js
@@ -4,7 +4,7 @@ import classnames from 'classnames';
 import { FormElement } from './../Form';
 import Icon from './../Icon';
 import { Button } from './../Button';
-import { DropdownMenu, DropdownMenuItem } from './../DropdownMenu';
+import { DropdownMenu, DropdownMenuItem } from './../Dropdown';
 import { uuid, isElInChildren } from './../utils';
 
 export default class Picklist extends Component {
