+++ lib/components/Tabs/Tabs.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
 import { registerStyle } from './../utils';
-import { DropdownButton } from './../DropdownMenu';
+import { DropdownButton } from './../Dropdown';
 
 /**
  *
